refactor: change project structure; move all organization related files to organization folder
