+++ backend/src/main/java/de/neuefische/backend/api/service/ArbeitsagenturApiService.java
@@ -5,8 +5,8 @@ import de.neuefische.backend.api.dto.ApiResponseCourseOffer;
 import de.neuefische.backend.api.dto.ApiResponseDetails;
 import de.neuefische.backend.api.dto.ApiResponseOrganization;
 import de.neuefische.backend.api.exception.ApiResponseException;
-import de.neuefische.backend.model.Organization;
-import de.neuefische.backend.service.IdService;
+import de.neuefische.backend.organization.IdService;
+import de.neuefische.backend.organization.model.Organization;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
