+++ backend/src/main/java/de/neuefische/backend/organization/IdService.java
@@ -1,4 +1,4 @@
-package de.neuefische.backend.service;
+package de.neuefische.backend.organization;
 
 import org.springframework.stereotype.Service;
