+++ backend/src/main/java/de/neuefische/backend/organization/model/Review.java
@@ -1,4 +1,4 @@
-package de.neuefische.backend.model;
+package de.neuefische.backend.organization.model;
 
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
