+++ backend/src/test/java/de/neuefische/backend/organization/OrganizationControllerIntegrationTest.java
@@ -1,8 +1,7 @@
-package de.neuefische.backend.controller;
+package de.neuefische.backend.organization;
 
 import de.neuefische.backend.api.service.ArbeitsagenturApiService;
-import de.neuefische.backend.model.Organization;
-import de.neuefische.backend.repository.OrganizationRepository;
+import de.neuefische.backend.organization.model.Organization;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
