+++ backend/src/test/java/de/neuefische/backend/organization/model/OrganizationTest.java
@@ -1,4 +1,4 @@
-package de.neuefische.backend.model;
+package de.neuefische.backend.organization.model;
 
 import org.junit.jupiter.api.Test;
